chore: refactor entry into deployment and runtime
RuntimeEntry has no startup value, and omits functions required
to be serpent compatible.
